breaks down osr api to separate files
