+++ filenames.gypi
@@ -206,6 +206,10 @@
       'atom/browser/native_window_mac.h',
       'atom/browser/native_window_mac.mm',
       'atom/browser/native_window_observer.h',
+      'atom/browser/osr_web_contents_view.cc',
+      'atom/browser/osr_web_contents_view.h',
+      'atom/browser/osr_output_device.cc',
+      'atom/browser/osr_output_device.h',
       'atom/browser/osr_window.cc',
       'atom/browser/osr_window.h',
       'atom/browser/osr_window_win.cc',
